+++ compact/day-of-week/jullian/stanislav_mikhailov.rs
@@ -193,16 +193,16 @@ fn day_of_week(year: u128, month: u8, day: u8) -> &'static str {
     let last_year: u128 = year - 1;
 
     let sum_leap_years_julian: u128 = last_year / 4;
-    let shift_week_day_in_jan: u8 = ((last_year + sum_leap_years_julian) % REPEAT_WEAK_DAY_CYCLE as u16 as u128) as u8;
+    let shift_week_day_in_jan: u8 = ((last_year + sum_leap_years_julian) % REPEAT_WEAK_DAY_CYCLE as u128) as u8;
 
     let shift_week_day_in_month_and_day: u8;
 
     if !is_leap_year_julian(year) {
-        shift_week_day_in_month_and_day = (BASE[(month - 1) as usize] + day) % REPEAT_WEAK_DAY_CYCLE as u16;
-        return week_day((shift_week_day_in_jan + shift_week_day_in_month_and_day) % REPEAT_WEAK_DAY_CYCLE as u16);
+        shift_week_day_in_month_and_day = (BASE[(month - 1) as usize] + day) % REPEAT_WEAK_DAY_CYCLE;
+        return week_day((shift_week_day_in_jan + shift_week_day_in_month_and_day) % REPEAT_WEAK_DAY_CYCLE);
     } else if is_leap_year_julian(year) {
-        shift_week_day_in_month_and_day = (LEAP[(month - 1) as usize] + day) % REPEAT_WEAK_DAY_CYCLE as u16;
-        return week_day((shift_week_day_in_jan + shift_week_day_in_month_and_day) % REPEAT_WEAK_DAY_CYCLE as u16);
+        shift_week_day_in_month_and_day = (LEAP[(month - 1) as usize] + day) % REPEAT_WEAK_DAY_CYCLE;
+        return week_day((shift_week_day_in_jan + shift_week_day_in_month_and_day) % REPEAT_WEAK_DAY_CYCLE);
     }
 
     panic!("Unkonwn error");
@@ -240,9 +240,9 @@ fn day_of_week(year: u128, month: u8, day: u8) -> &'static str {
     let last_year: u128 = year - 1;
 
     if !is_leap_year_julian(year) {
-        return week_day(((last_year + (year / 4) + BASE[(month - 1) as usize] as u128 + day as u128) % REPEAT_WEAK_DAY_CYCLE as u16 as u128) as u8);
+        return week_day(((last_year + (year / 4) + BASE[(month - 1) as usize] as u128 + day as u128) % REPEAT_WEAK_DAY_CYCLE as u128) as u8);
     } else if is_leap_year_julian(year) {
-        return week_day(((last_year + (year / 4) + LEAP[(month - 1) as usize] as u128 + day as u128) % REPEAT_WEAK_DAY_CYCLE as u16 as u128) as u8);
+        return week_day(((last_year + (year / 4) + LEAP[(month - 1) as usize] as u128 + day as u128) % REPEAT_WEAK_DAY_CYCLE as u128) as u8);
     }
 
     panic!("Unkonwn error");
